In work. Pickling won't work because it doesn't handle SWIG objects.
