+++ phaseflow/__init__.py
@@ -1,1 +1,1 @@
-from . import core, pure, pure_with_constant_properties, benchmarks
+from . import state, simulation, octadecane, octadecane_benchmarks
